requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
passlib==1.7.4  
bcrypt==4.1.3   
python-jose[cryptography]
python-multipart
scikit-learn
pandas
firebase-admin
mlxtend
matplotlib
seaborn
apscheduler
pytz
colorama